requirements.txt
langchain>=0.1.0
langchain-google-genai>=1.0.0
langgraph>=0.0.26
chromadb>=0.4.0
python-dotenv>=1.0.0
fastapi>=0.100.0
uvicorn[standard]>=0.24.0
requests>=2.31.0
streamlit>=1.28.0
rank_bm25>=0.2.2
plotly>=5.18.0
pandas>=2.0.0
langchain-openai

# For semantic similarity and embeddings
sentence-transformers==2.2.2

# For NLP-based metrics (Named Entity Recognition, etc.)
spacy==3.7.2

# spaCy language model (run this after installing spacy):
# python -m spacy download en_core_web_sm

# For numerical operations (if not already installed)
numpy>=1.24.0

# For better text processing
nltk>=3.8.1

matplotlib